Refactor Changes
