+++ src/container/GeoWebInterface/components/GeoWebContent/styles.css
@@ -1,0 +1,20 @@
+.switch-div {
+    position: relative;
+    margin-top: 2%;
+}
+
+.switch-left {
+    position: absolute;
+    top: 12%;
+    left: 14%;
+    color: white;
+}
+
+.switch-right {
+    position: absolute;
+    float: right;
+    right: 20%;
+    top: 12%;
+    color: white;
+}
+
