+++ src/container/GeoWebInterface/components/GeoWebView/styles.css
@@ -1,0 +1,7 @@
+.gwIframe {
+    position: relative;
+    margin-top: 2%;
+    width: 100%;
+    height: calc(100vh - 200px);
+    overflow:hidden;
+}
